requirements.txt
asgiref==3.8.1
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.0.3
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
filelock==3.13.4
Flask==3.0.2
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysqlclient==2.2.4
numpy==1.26.4
oauthlib==3.2.2
platformdirs==4.2.0
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.3
virtualenvwrapper-win==1.2.7
Werkzeug==3.0.1
